Applied enhancements to "chess" package structure
